renames UI subdirectory to styles
