+++ components/Rating.tsx
@@ -1,6 +1,6 @@
 import { TypographyProps, typography } from "styled-system";
 
-import Flex from "./styled/Flex";
+import Flex from "./styles/Flex";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faStar } from "@fortawesome/free-solid-svg-icons";
 import styled from "styled-components";
