+++ components/ShowHeader.tsx
@@ -7,10 +7,10 @@ import {
   typography,
 } from "styled-system";
 
-import Flex from "./styled/Flex";
+import Flex from "./styles/Flex";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import Link from "next/link";
-import Padding from "./styled/Padding";
+import Padding from "./styles/Padding";
 import Rating from "./Rating";
 import { faStar } from "@fortawesome/free-solid-svg-icons";
 import styled from "styled-components";
